refactor(): Convertido projeto de tsx para tsx e removido partes do template nao utilizado
