+++ app/(tabs)/_layout.jsx
@@ -1,0 +1,28 @@
+import { Ionicons } from '@expo/vector-icons';
+import { Tabs } from 'expo-router';
+import React from 'react';
+
+import { useColorScheme } from '@/components/useColorScheme';
+import Colors from '@/constants/Colors';
+
+export default function TabLayout() {
+  const colorScheme = useColorScheme() ?? 'light';
+
+  return (
+    <Tabs
+      screenOptions={{
+        tabBarActiveTintColor: Colors[colorScheme].tint,
+        headerShown: true,
+      }}>
+      <Tabs.Screen
+        name="index"
+        options={{
+          title: 'Mapa',
+          tabBarIcon: ({ color }) => (
+            <Ionicons name="map-outline" size={28} color={color} />
+          ),
+        }}
+      />
+    </Tabs>
+  );
+}
