+++ app/(tabs)/index.jsx
@@ -1,0 +1,21 @@
+import { StyleSheet } from 'react-native';
+
+export default function TabOneScreen() {
+}
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    alignItems: 'center',
+    justifyContent: 'center',
+  },
+  title: {
+    fontSize: 20,
+    fontWeight: 'bold',
+  },
+  separator: {
+    marginVertical: 30,
+    height: 1,
+    width: '80%',
+  },
+});
