+++ tsconfig.json.bak
@@ -1,0 +1,17 @@
+{
+  "extends": "expo/tsconfig.base",
+  "compilerOptions": {
+    "strict": true,
+    "paths": {
+      "@/*": [
+        "./*"
+      ]
+    }
+  },
+  "include": [
+    "**/*.ts",
+    "**/*.tsx",
+    ".expo/types/**/*.ts",
+    "expo-env.d.ts"
+, "app/(tabs)/_layout.jsx", "app/(tabs)/index.jsx", "app/(tabs)/settings.jsx", "app/(tabs)/two.jsx", "app/_layout.jsx", "app/+not-found.jsx", "app/index.jsx", "app/modal.jsx", "components/EditScreenInfo.jsx", "components/ExternalLink.jsx", "components/StyledText.jsx", "components/Themed.jsx", "components/useClientOnlyValue.js", "components/useClientOnlyValue.web.js", "components/useColorScheme.js", "components/useColorScheme.web.js", "constants/Colors.js"  ]
+}
